+++ lib/src/clients/DomainClient.dart
@@ -48,9 +48,10 @@ class DomainClient extends AbstractDomainRobotClient {
     var payloadAsString = json.encode(payload.toJson());
     Map<String, dynamic> body;
     try {
-      body = await HttpUtils.putForJson(
-          '$baseUrl/domain/${payload.name}', payloadAsString,
-          queryParameters: queryParameters, headers: headers);
+      body = await HttpUtils.putForJson('$baseUrl/domain/${payload.name}',
+          body: payloadAsString,
+          queryParameters: queryParameters,
+          headers: headers);
     } catch (e) {
       AbstractDomainRobotClient.handleException(e);
     }